backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-api-python-client==2.123.0
youtube-transcript-api==0.6.2
sentence-transformers==3.3.1
scikit-learn==1.4.2
torch==2.3.1
numpy==2.0.2
httpx==0.27.0
pydantic==2.6.4
